Change hourglass dim check test to input expected pattern
* Rename model factories utils module
